# app/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Client:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
      tg_id:
        type: integer
      tg_nickname:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления GoGRAFF
  title: GoGRAFF API
  version: "1.0"
paths:
...
  /clients:
    get:
      description: Возвращает список всех клиентов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить всех клиентов
      tags:
      - Клиенты
    post:
      consumes:
      - application/json
      description: Создает нового клиента
      parameters:
      - description: Данные клиента
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать клиента
      tags:
      - Клиенты
  /clients/{id}:
    delete:
      description: Удаляет клиента по ID
      parameters:
      - description: ID клиента
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удалить клиента
      tags:
      - Клиенты
    get:
      description: Возвращает данные клиента по ID
      parameters:
      - description: ID клиента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Клиент не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить клиента
      tags:
      - Клиенты
    put:
      consumes:
      - application/json
      description: Обновляет данные клиента по ID
      parameters:
      - description: ID клиента
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные клиента
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Клиент не найден
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновить клиента
      tags:
      - Клиенты
  /clients/check:
    get:
      description: Проверяет, существует ли клиент с заданными контактными данными
      parameters:
      - description: Номер телефона клиента
        in: query
        name: phone_number
        type: string
      - description: Telegram ID клиента
        in: query
        name: tg_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Результат проверки
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Проверить существование клиента
      tags:
      - Клиенты
  /clients/filter:
    get:
      description: Возвращает список клиентов по имени
      parameters:
      - description: Имя клиента для фильтрации
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Client'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Фильтрация клиентов по имени
      tags:
      - Клиенты
  /clients/quick_add:
    post:
      consumes:
      - application/json
      description: Создает клиента с минимальными данными (например, через Telegram
        и/или номер телефона)
      parameters:
      - description: Минимальные данные клиента
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Клиент уже существует
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Быстрое добавление клиента
      tags:
      - Клиенты
  /clients/search:
    get:
      description: Ищет клиента по Email или номеру телефона
      parameters:
      - description: Email клиента
        in: query
        name: email
        type: string
      - description: Номер телефона клиента
        in: query
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Клиент не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Найти клиента по контактным данным
      tags:
      - Клиенты
  /clients/telegram/{tg_id}:
    get:
      description: Возвращает данные клиента по Telegram ID
      parameters:
      - description: Telegram ID клиента
        in: path
        name: tg_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Client'
        "404":
          description: Клиент не найден
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить клиента по Telegram ID
      tags:
      - Клиенты
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"